+++ tests/test_report.py
@@ -1,5 +1,3 @@
-import sys
-
 import pytest
 
 from . import v
@@ -47,8 +45,6 @@ def test_item_report(check_report):
 {filename}:11: unreachable code after 'return' (100% confidence)
 {filename}:13: unused property 'myprop' (60% confidence)
 """
-    if sys.version_info >= (3, 8):
-        expected = expected.replace("{filename}:13:", "{filename}:14:")
     check_report(mock_code, expected)
 
 
@@ -62,6 +58,4 @@ foobar  # unused variable ({filename}:9)
 # unreachable code after 'return' ({filename}:11)
 _.myprop  # unused property ({filename}:13)
 """
-    if sys.version_info >= (3, 8):
-        expected = expected.replace("{filename}:13)", "{filename}:14)")
     check_report(mock_code, expected, make_whitelist=True)